Do not report old or duplicate pod termination errors
If a Pod is updated k8s-sentry receives an update event. When we examined the
pod status and saw a termination record with a non-zero exit code we always
reported that as an error. This resulted in the same error being reported
again and again if the pod status was updated for other reasons. To fix that
we now apply two extra rules: 1) ignore old termination records (over 5
seconds), and keep a small LRU cache of terminations we have seen before so we
can detect duplicates.