feat: Rearrange the pod observation
The podRailed case seemed pretty useless to me as the message and reason
was always empty so I got a lot of meaningless events. I also think that
any Pod-Failure that is not due to failed containers will show up as
events and get better reported as such.

Another disadvantage of the former implementation was, that it reported
only ever one container failure per pod.